pyproject.toml
[project]
name = "aww-agent"
version = "0.1.0"
description = "Lightweight agent framework"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.76.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "rich>=14.2.0",
    "ruff>=0.14.13",
]
